requirements.txt
pandas~=1.3.4
SQLAlchemy~=1.4.27
python-dotenv~=0.19.2
httpx~=0.21.1
pytomlpp~=1.0.9
tzlocal~=4.1
numpy~=1.21.4
XlsxWriter~=3.0.2
websockets~=9.1
streamlit~=1.2.0
ta~=0.8
matplotlib~=3.5.1
python-dateutil~=2.8.2
pytz~=2021.3